Fix dtypes for metrics calculation

+++ src/csbm/models/images.py
@@ -503,7 +503,7 @@ class ImageD3PM(nn.Module):
         for t in reversed(range(1, self.num_timesteps + 2)):
             t = torch.tensor([t] * x.shape[0], device=self.device)
             x = self.markov_sample(x, t, prior)
-        return x
+        return x.to(dtype=torch.uint8)
     
     @torch.no_grad()
     def sample_trajectory(self, x: torch.Tensor, prior: Prior) -> torch.Tensor:
